fix(interactive.imagetool): various fixes related to manager
This fix incorporates many changes to the ImageTool and ImageTool Manager.

First, the archiving function of the manager now works properly, and tries to clear memory eagerly.

When opening data from a file using the GUI, the name of the file will now be displayed in the title bar of the ImageTool. This file name is also propagated to the name displayed in the manager.

Furthermore, the archiving and show/hide functionality of the manager has been updated to restore the window geometry automatically. When the user shows or unarchives a hidden or archived window, the previous position of the window  is restored.

Some icons and the layout of the manager has been modified, and tooltips has been added to the buttons.

Also, some unexpected behavior regarding linking has been resolved.