refactor: change temperature attribute name
Changes `temp_sample` to `sample_temp` for all data loaders and analysis code.

BREAKING CHANGE: The default attribute name for the sample temperature is changed to `sample_temp` from `temp_sample`. This will unfortunately break a lot of code that relies on the key `temp_sample`, but will be easy to refactor with find and replace.