refactor(imagetool): disable manager by default
This update modifies `itool` and `qshow` so that tools are opened in the manager only when `use_manager=True` is explicitly specified.

To address the inconvenience, a new `Move to Manager` action with keyboard shortcut `Ctrl+Shift+M` has been added to the `File` menu of ImageTool windows opened outside the manager when the manager is running.