feat(imagetool): enable directly opening tools with cropped data
A new feature has been added to the imagetool that allows users to open the tool with cropped data directly. When holding down the Alt key (Option on Mac) while in the right-click menu of the image, relevant menus will automatically work with the data cropped to the currently visible range.