fix(io.plugins.merlin): take energy step from data attributes
Fixed and swept mode scans have random numerical errors in the start & step of the energy axis. The error of the step accumulates over the scan, effectively resulting in a tiny renormalization of the energy axis. We choose to ignore the energy axis information contained in the wave, and take the step from the attributes instead.