feat(imagetool.manager): enhance data replacement logic (#153)
The `replace` argument for `itool` now supports negative indices to replace the newest ImageTool window, and allows for replacing the next index if it has not yet been created. For example, if you have 2 windows open (indices 0 and 1), you can use `replace=-1` to replace the newest window (index 1) or `replace=2` to create a new window if it does not exist yet.