fix(imagetool): make ImageTool work with dask arrays
Passing dask arrays to ImageTool used to load the entire array into memory. This commit fixes that by ensuring that the data is kept as a dask array and is indexed lazily. However, this makes the GUI much slower to respond, so it is recommended to use ImageTool with in-memory arrays when possible.