feat(manager): enhance integration with Jupyter notebooks
Adds a new magic command `%watch` to monitor and sync data between ImageTool manager and Jupyter notebooks. This feature allows users to easily update data in the notebook when changes are made in the ImageTool window and vice versa, facilitating a smoother workflow for data analysis and visualization.

Also adds a new function, `erlab.interactive.imagetool.manager.fetch` which allows users to fetch a copy of the data in a specific ImageTool window in the manager by its index.