feat(imagetool): auto-compute small dask arrays
When loading data into the imagetool, if the data is a dask array and its size is below a certain threshold (configurable via `erlab.interactive.options["io/compute_threshold"]` or the settings GUI), the data will be automatically computed into memory. Default threshold is 2048 MB.

Additionally, adds a "Load Into Memory" action to the `File` menu in ImageTool only visible for chunked arrays which calls `.compute()` on the data when triggered.