perf(imagetool): greatly improve performance for dask-based data (#197)
Previously, indexing/slicing/averaging was done for all plots individually. This commit changes the behavior for dask-based data so that dask.compute is now called only once for a given update, which greatly improves performance when working with dask. Regular data retains the original behavior, but should benefit from additional optimizations such as improved signal handling when moving multiple cursors simultaneously and ignored image updates for plots that are hidden.