feat(analysis.gold): allow fermi edge fitting functions to be used with any angle dimension and higher dimensional data
Adds an 'along' parameter to the functions like `edge`, `poly`, and `correct_with_edge` to specify which dimension to use for fitting. Previously, these functions assumed an 'alpha' dimension, limiting their applicability.

With this change, the functions can also fit higher dimensional data; for instance, a 3D DataArray containing hν-alpha-eV data can be passed directly to `edge` or `poly` to get a multidimensional fit.

Dask parallelization is also automatically handled for chunked inputs, as long as the data is not chunked along the energy dimension. Previously, passing a dask array to `edge` used to fail miserably.